unit test reorg and expanded coverage
