+++ test/url_template_test.dart
@@ -6,20 +6,33 @@ import 'package:route_hierarchical/url_matcher.dart';
 
 main() {
   group('UrlTemplate', () {
+    test('toString should return stringified url pattern', () {
+      var tmpl = new UrlTemplate('/foo/bar:baz/aux');
+      expect(tmpl.urlParameterNames, equals(['baz']));
+      expect(
+          tmpl.toString(),
+          equals(
+              'UrlTemplate(JSRegExp: pattern=^/foo/bar([^/?]+)/aux flags=)'));
+    });
+
     test('should work with simple templates', () {
       var tmpl = new UrlTemplate('/foo/bar:baz/aux');
+      expect(tmpl.urlParameterNames, equals(['baz']));
       expect(tmpl.match('/foo/bar123/aux'),
           new UrlMatch('/foo/bar123/aux', '', {'baz': '123'}));
 
       tmpl = new UrlTemplate('/foo/:bar');
+      expect(tmpl.urlParameterNames, equals(['bar']));
       expect(
           tmpl.match('/foo/123'), new UrlMatch('/foo/123', '', {'bar': '123'}));
 
       tmpl = new UrlTemplate('/:foo/bar');
+      expect(tmpl.urlParameterNames, equals(['foo']));
       expect(
           tmpl.match('/123/bar'), new UrlMatch('/123/bar', '', {'foo': '123'}));
 
       tmpl = new UrlTemplate('/user/:userId/article/:articleId/view');
+      expect(tmpl.urlParameterNames, equals(['userId', 'articleId']));
       UrlMatch params =
           tmpl.match('/user/jsmith/article/1234/view/someotherstuff');
       expect(
@@ -31,6 +44,7 @@ main() {
       expect(params, isNull);
 
       tmpl = new UrlTemplate(r'/foo/:bar$123/aux');
+      expect(tmpl.urlParameterNames, equals(['bar']));
       expect(tmpl.match(r'/foo/123$123/aux'),
           new UrlMatch(r'/foo/123$123/aux', '', {'bar': '123'}));
     });
@@ -40,6 +54,7 @@ main() {
       expect(tmpl.match(r'\^\|+[]{}()'), new UrlMatch(r'\^\|+[]{}()', '', {}));
 
       tmpl = new UrlTemplate(r'/:foo/^\|+[]{}()');
+      expect(tmpl.urlParameterNames, equals(['foo']));
       expect(tmpl.match(r'/123/^\|+[]{}()'),
           new UrlMatch(r'/123/^\|+[]{}()', '', {'foo': '123'}));
     });
@@ -57,33 +72,40 @@ main() {
 
     test('should reverse', () {
       var tmpl = new UrlTemplate('/:a/:b/:c');
+      expect(tmpl.urlParameterNames, equals(['a', 'b', 'c']));
       expect(tmpl.reverse(), '/null/null/null');
       expect(tmpl.reverse(parameters: {'a': 'foo', 'b': 'bar', 'c': 'baz'}),
           '/foo/bar/baz');
 
       tmpl = new UrlTemplate(':a/bar/baz');
+      expect(tmpl.urlParameterNames, equals(['a']));
       expect(tmpl.reverse(), 'null/bar/baz');
       expect(tmpl.reverse(parameters: {'a': '/foo',}), '/foo/bar/baz');
 
       tmpl = new UrlTemplate('/foo/bar/:c');
+      expect(tmpl.urlParameterNames, equals(['c']));
       expect(tmpl.reverse(), '/foo/bar/null');
       expect(tmpl.reverse(parameters: {'c': 'baz',}), '/foo/bar/baz');
 
       tmpl = new UrlTemplate('/foo/bar/:c');
+      expect(tmpl.urlParameterNames, equals(['c']));
       expect(tmpl.reverse(tail: '/tail', parameters: {'c': 'baz',}),
           '/foo/bar/baz/tail');
     });
 
     test('should conditionally allow slashes in parameters', () {
       var tmpl = new UrlTemplate('/foo/:bar');
+      expect(tmpl.urlParameterNames, equals(['bar']));
       expect(tmpl.match('/foo/123/456'),
           new UrlMatch('/foo/123', '/456', {'bar': '123'}));
 
       tmpl = new UrlTemplate('/foo/:bar*');
+      expect(tmpl.urlParameterNames, equals(['bar*']));
       expect(tmpl.match('/foo/123/456'),
           new UrlMatch('/foo/123/456', '', {'bar*': '123/456'}));
 
       tmpl = new UrlTemplate('/foo/:bar*/baz');
+      expect(tmpl.urlParameterNames, equals(['bar*']));
       expect(tmpl.match('/foo/123/456/baz'),
           new UrlMatch('/foo/123/456/baz', '', {'bar*': '123/456'}));
     });
